% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_BoolFun.R
\name{BoolBioNet_BoolFun}
\alias{BoolBioNet_BoolFun}
\title{Explore the features of Boolean functions in biological networks}
\usage{
BoolBioNet_BoolFun(OneBioNet, Type = c("rough", "detail"))
}
\arguments{
\item{OneBioNet}{A four-element formatted list that records information of a Boolean network.}

\item{Type}{Which analysis method is applied, \code{rough} or \code{detail}?}
}
\value{
A \code{data.frame} summarizing the count of functions by type when \code{Type} is set to \code{rough} (or a \code{data.frame} listing the specific instances of each function within systems when \code{Type} is set to \code{detail}).
}
\description{
This function computes and returns the category, sensitivity, effectiveness, and complexity of each Boolean function within a given Boolean network.
}
\details{
The function can identify various function types, including canalizing, threshold, clique, dominating, effective, monotone, signed, and spin-like functions. Definitions of these functions are available in the documentation of \link{BoolFun_Type}.
Note that external factors are excluded from the analysis; only nodes with in-degrees between 2 and 15 (inclusive) are considered, while other cases return \code{NA}. When \code{Type} is set to \code{rough}, the function returns a \code{data.frame} containing summary statistics for each "in-degree" cluster. When \code{Type} is \code{detail}, the returned \code{data.frame} contains individual cases of the Boolean function.
}
\examples{
# Analysis of one real Boolean genetic network.
# BoolBioNet_BoolFun(BoolGRN_CellCollective$c_1778,"rough");
# Return a data.frame:
#                 Number Canalized Clique Monotone Signed Threshold
# InDegree_1           1        NA     NA       NA     NA        NA
# InDegree_2           3         3      3        1      3         3
# InDegree_3           3         3      2        0      3         3
#                             ... ...
BoolBioNet_BoolFun(BoolGRN_CellCollective$c_1778,"rough")

# BoolBioNet_BoolFun(BoolGRN_CellCollective$c_1778,"detail");
# Return a data.frame:
#            Input Canalized Clique Monotone Signed Threshold Sensitivity   ... 
# ADD            5      TRUE  FALSE    FALSE   TRUE      TRUE   1.3125000   ...
# ATM            3      TRUE  FALSE    FALSE   TRUE      TRUE   1.2500000   ...
# ATR            4      TRUE  FALSE    FALSE   TRUE      TRUE   1.2500000   ...
#                                ... ...
BoolBioNet_BoolFun(BoolGRN_CellCollective$c_1778,"detail")

}
