% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_CoreDyn.R
\name{BoolBioNet_CoreDyn}
\alias{BoolBioNet_CoreDyn}
\title{Analyze core dynamic of Boolean networks}
\usage{
BoolBioNet_CoreDyn(
  RealBioNet,
  AnalysisMod = c("coredyn", "scaling"),
  ExternalSelfLoop = TRUE,
  Controller = NULL,
  ControllerValue = NULL,
  ExternalNode = NULL,
  NodeAttri = FALSE,
  ResidualNet = FALSE,
  Times = 100L
)
}
\arguments{
\item{RealBioNet}{a 4-element formatted list, record biological network information.}

\item{AnalysisMod}{string, \code{coredyn} for core dynamic analysis, \code{scaling} for searching all engaged nodes.}

\item{ExternalSelfLoop}{logical, should convert all external inputs as self-loop nodes? (Default: \code{TRUE})}

\item{Controller}{integer or character vector, denote which nodes should be controlled. (Default: \code{NULL})}

\item{ControllerValue}{Boolean vector, controlled nodes' values. (Default: \code{NULL})}

\item{ExternalNode}{Boolean vector, set non input nodes' state (Default: \code{NULL}, random setting, see details)}

\item{NodeAttri}{logical, should return node's attribute? (Default: \code{FALSE}, ruturn value: \code{NA})}

\item{ResidualNet}{logical, should return residual network? (Default: \code{FALSE}, ruturn value: \code{NA})}

\item{Times}{an positive integer, recursive number of analysis; not recommended for users to change it.}
}
\value{
List[4], [[1]]Overview: [1] size stable node, [2] size of invalid node, [3] size of valid nodes, 
[4] size of external component, [5] size of terminal component
[[2]]Node_S01U: Node's attribute 1(\code{IntegerVector} or \code{NA}), 0-stable (as 1), 1-stable (as 2), unstable (as 3).
[[3]]Node_SUE: Node's attribute 2(\code{IntegerVector} or \code{NA}), stable (as 1), useless (as 0), engaged (as -1).
[[4]]ResidualNetwork: residual network (\code{List[4]} or \code{NA}), still contains all nodes.
}
\description{
This analysis only emphasizes the static feature of the network.
The function recursively analyzes core dynamic components (\code{coredyn}) or 
engaged nodes (\code{scaling}) that contribute to systematic dynamic feature. 
These two indicators consider coupled feedforward loop or not, respectively.
}
\details{
Some nodes within bio-networks have no input, called as free-nodes.
If \code{ExternalSelfLoop} is \code{TRUE}, all free-nodes are added a self-loop
circuits. When it is \code{FALSE}, free-nodes depend on following parameters.
\code{ExternalNode} configure all free-nodes. If \code{ExternalNode} is \code{NULL},
they can be set via series random Boolean values (0/1). If provided, function only use the [1,N_FreeNode] 
values to set; when its length shorter than N_FreeNode, function return an error message and stop. 

\code{Controller} denotes manual setting of controlling nodes. Node's indexes or names are both acceptable. Note
that function does not check index's correctness. Please ensure the correct indexes. \code{ControllerValue} denotes 
corresponding values of controlled genes. If not provided (\code{NULL}), they are generated randomly. If provided,
only the [1,N_Control] are utilized; shorter than \code{N_Control} return an error message and stop. One gene is
both included in \code{Controller} and \code{ExternalNode}, the configurated value of the former covers the latter one. 

\code{Residual networks} still contain all nodes within original systems for comparison. Stable nodes, useless 
nodes, and corresponding egdes would be removed and cut-off. Remaing nodes, edges, and Boolan functions
are engaged in terminal dynamic behaviors.
}
\examples{
# Analysis of core dynamic components of a Boolean genetic network.
# BoolBioNet_CoreDyn(BoolGRN_CellCollective$c_11863, 
#   ExternalSelfLoop=TRUE);# Treat external nodes as self-loop!
# Return [[1]] [1] 0, 28, 23, 0, 0
#        [[2]] ~ [[4]] NA
BoolBioNet_CoreDyn(BoolGRN_CellCollective$c_11863, 
  ExternalSelfLoop=TRUE);

# set.seed(1234);
# BoolBioNet_CoreDyn(BoolGRN_CellCollective$c_11863, 
#   ExternalSelfLoop=FALSE);# Treat free-node as fixed nodes!
# Return [[1]] [1] 51, 0, 0, 2, 0
#        [[2]] ~ [[4]] NA
set.seed(1234)
BoolBioNet_CoreDyn(BoolGRN_CellCollective$c_11863, 
  ExternalSelfLoop=FALSE)

}
