% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolFun_Complexity.R
\name{BoolFun_Complexity}
\alias{BoolFun_Complexity}
\title{Calculate the complexity of a Boolean function}
\usage{
BoolFun_Complexity(Bit_Vec)
}
\arguments{
\item{Bit_Vec}{a bool (or 0/1-integral) vector, represents a truth table of Boolean function.}
}
\value{
double, complexity of a Boolean function
}
\description{
The analysis depends on average of prime implicants in Boolean 
function and its complementary one, based on Boolean Quine-McCluskey method.
}
\examples{
# Calculate the complexity of a Boolean function \code{c(0,0,1,0,0,1,1,0)}.
# BoolFun_Complexity(c(0,0,1,0,0,1,1,0))
# {(*10)+(101) ==> 1 (00*)+(*00)+(*11) ==> 0 } -> 2.5
# Return 5
BoolFun_Complexity(c(0,0,1,0,0,1,1,0))
}
