% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolFun_Expr2MapTable.R
\name{BoolFun_Expr2MapTable}
\alias{BoolFun_Expr2MapTable}
\title{Convert Boolean expressions into mapping tables}
\usage{
BoolFun_Expr2MapTable(BoolExpChr, TableForm = FALSE)
}
\arguments{
\item{BoolExpChr}{A string representing a valid Boolean expression.}

\item{TableForm}{A logical value. Should return a table format? (\code{TRUE}: 
matrix; \code{FALSE}: Boolean vector)}
}
\value{
A Boolean vector or a mapping table.
}
\description{
The objective is to uncover the relationships between the mapping
results of Boolean functions and their inputs, with implementation based on 
regular expressions.
}
\details{
Note that the function does not verify the correctness of the Boolean expression.
}
\examples{
# Convert \code{X = ((NOT A AND B)) OR C}.
# BoolFun_Expr2MapTable("X = ((NOT A AND B)) OR C");
# Return a BooleanVector c(FALSE, TRUE, TRUE, TRUE, FALSE, TRUE, FALSE, TRUE).
boolexp="X = ((NOT A AND B)) OR C"
BoolFun_Expr2MapTable(boolexp)

# Show-only \code{Y = (NOT A_a OR NOT B) AND (cc OR dd)}.
# BoolFun_Expr2MapTable("Y = (NOT A_a OR NOT B) AND (cc OR dd)", TRUE)
# Return a truth table: 
#      A_a  B  cc  dd  Y
# [0]    0  0   0   0  0
# [1]    0  0   0   1  1
# [2]    0  0   1   0  1
#           ... ... 
# Last column is c(0,1,1,1, 0,1,1,1, 0,1,1,1, 0,0,0,0)
boolexp="Y = (NOT A_a OR NOT B) AND (cc OR dd)"
BoolFun_Expr2MapTable(boolexp, TRUE)

}
