% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_Complexity}
\alias{c_BF_Complexity}
\title{Calculate the complexity of a Boolean function}
\usage{
c_BF_Complexity(boolfuns, k)
}
\arguments{
\item{boolfuns}{[\code{IntegerVector &}] the Boolean function.}

\item{k}{[\code{int}] the number of input variables of the function.}
}
\description{
C++ prototype function of \link{BoolFun_Complexity}.
}
