% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_EffectiveEdges}
\alias{c_BF_EffectiveEdges}
\title{Calculate the effective edges of a Boolean function}
\usage{
c_BF_EffectiveEdges(boolfuns, k)
}
\arguments{
\item{boolfuns}{[\code{IntegerVector &}] the Boolean function.}

\item{k}{[\code{int}] the number of input variables of the function.}
}
\description{
C++ prototype function of \link{BoolFun_Effectiveness}.
}
