% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_Generator}
\alias{c_BF_Generator}
\title{C++ prototype function for BoolFun_Generator()}
\usage{
c_BF_Generator(Leixing, k, bias, Vars)
}
\arguments{
\item{Leixing, }{function's type.}

\item{k, }{input variable size.}

\item{bias, }{function's bias.}

\item{Vars, }{some configuration parameters.}
}
\description{
C++ prototype function for BoolFun_Generator()
}
\details{
Please see document of \link{BoolFun_Generator}
}
