% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_QuineMcCluskey}
\alias{c_BF_QuineMcCluskey}
\title{C++ prototype function for BoolFun_QMForm()}
\usage{
c_BF_QuineMcCluskey(boolfunss, k, VarsName)
}
\arguments{
\item{boolfunss, }{bool function logical vector.}

\item{k, }{input variable size.}

\item{VarsName, }{variables' names.}
}
\description{
C++ prototype function for BoolFun_QMForm()
}
\details{
Please see document of \link{BoolFun_QMForm}
}
