% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_Sensitivity}
\alias{c_BF_Sensitivity}
\title{C++ prototype function for BoolFun_Sensitivity()}
\usage{
c_BF_Sensitivity(boolfunss, k)
}
\arguments{
\item{boolfunss, }{bool function logical vector.}

\item{k, }{input variable size.}
}
\description{
C++ prototype function for BoolFun_Sensitivity()
}
\details{
Please see document of \link{BoolFun_Sensitivity}
}
