% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_FrameTruthTable}
\alias{c_FrameTruthTable}
\title{Enumerate all possible input vectors}
\usage{
c_FrameTruthTable(VarNum)
}
\arguments{
\item{VarNum}{[\code{int}] the number of input variables of the function}
}
\description{
C++ prototype function for enumerate all possible input vectors
under the given length.
}
