% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulF_Effective}
\alias{c_MulF_Effective}
\title{Calculate the effetive edges loading of a multi-valued function}
\usage{
c_MulF_Effective(avec, k, L)
}
\arguments{
\item{avec}{[\code{IntegerVector}] the function.}

\item{k}{[\code{int}] the number of input variables.}

\item{L}{[\code{int}] the level of function.}
}
\value{
[\code{double}] the global effetive or redundant edges of given function.
}
\description{
C++ prototype function (1) of \link{MulVFun_Effectiveness}
}
