% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulF_EffectiveEdges}
\alias{c_MulF_EffectiveEdges}
\title{Calculate the effetive edges loading of a multi-valued function}
\usage{
c_MulF_EffectiveEdges(avec, k, L)
}
\arguments{
\item{avec}{[\code{IntegerVector}] the function.}

\item{k}{[\code{int}] the number of input variables.}

\item{L}{[\code{int}] the level of function.}
}
\value{
[\code{NumericVector}] the effetive or redundant value of each edge.
}
\description{
C++ prototype function (2) of \link{MulVFun_Effectiveness}
}
