% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulF_QuineMcCluskey}
\alias{c_MulF_QuineMcCluskey}
\title{Show the Quine-McCluskey form of a multi-valued function}
\usage{
c_MulF_QuineMcCluskey(avec, k, L)
}
\arguments{
\item{avec}{[\code{IntegerVector &}] that represents a mapping table of multi-valued function.}

\item{k}{[\code{int}] that denotes the number of input variables.}

\item{L}{[\code{int}] that represents the level of multi-valued system.}
}
\description{
C++ prototype function of \link{MulVFun_QMForm}
}
