% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulV2Bool_Bool2MulV}
\alias{c_MulV2Bool_Bool2MulV}
\title{C++ prototype function for MulV2Bool_Bool2MulV()}
\usage{
c_MulV2Bool_Bool2MulV(OriMapTab, k, L, Thresholds, b2m)
}
\arguments{
\item{OriMapTab, }{IntegerVector of original mapping table.}

\item{k, }{input variable size.}

\item{L, }{L-level discrete value.}

\item{Thresholds, }{IntegerVector of binary thresholds for multi-valued system.}

\item{b2m, }{Boolean value, Is from Boolean to Multi-valued system?}
}
\description{
C++ prototype function for MulV2Bool_Bool2MulV()
}
\details{
Please see document of \link{MulV2Bool_Bool2MulV}
}
