% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulVF_Sensitivity}
\alias{c_MulVF_Sensitivity}
\title{C++ prototype function for MulVFun_Sensitivity()}
\usage{
c_MulVF_Sensitivity(amulfunss, k, L, Lens)
}
\arguments{
\item{amulfunss, }{a multi-valued function vector.}

\item{k, }{input variable size.}

\item{L, }{L-level discrete value.}

\item{Lens, }{length of mapping table: L^k.}
}
\description{
C++ prototype function for MulVFun_Sensitivity()
}
\details{
Please see document of \link{MulVFun_Sensitivity}
}
