% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulVFun2Polynomial}
\alias{c_MulVFun2Polynomial}
\title{Convert a multi-valued function as polynomial like forms}
\usage{
c_MulVFun2Polynomial(VariableMat, MapTab, k, L)
}
\arguments{
\item{VariableMat}{[\code{IntegerMatrix &}] a matrix that record input vectors.}

\item{MapTab}{[\code{IntegerVector &}] the function.}

\item{k}{[\code{int}] the number of input variables of original function.}

\item{L}{[\code{int}] the level of original function.}
}
\description{
C++ prototype function of \link{MulVFun_Polynomial}
}
