% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_StrongConnectComponent}
\alias{c_StrongConnectComponent}
\title{Caluclate the network's strongly connected components}
\usage{
c_StrongConnectComponent(aRealNet, InD, OtD)
}
\arguments{
\item{aRealNet}{[\code{List}] a real/silico genetic network (See \link{BoolGRN_CellCollective}).}

\item{InD}{[\code{IntegerVector}] the vector of in-degrees.}

\item{OtD}{[\code{IntegerVector}] the vector of out-degrees.}
}
\description{
C++ prototype function of \link{BoolBioNet_StrConComp}
}
