% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigLatticePlot.R
\name{FigLatticePlot}
\alias{FigLatticePlot}
\title{Show the percolation cases}
\usage{
FigLatticePlot(
  Values = NULL,
  Length = NULL,
  Height = NULL,
  Type = c(4L, 3L, 6L),
  ColorFills = c("#efefef", "#EE312E", "#2156A6"),
  LineColor = "#111111",
  Linewidth = 0.7
)
}
\arguments{
\item{Values}{an IntegerVector, the results from \link{DNS_Percolation}}

\item{Length}{an integer, the lngth of lattice}

\item{Height}{an integer, the height of lattice}

\item{Type}{an integer, 4=square; 3=triangle; 6=hexagon}

\item{ColorFills}{a CharacterVector, [1] unstable, [2]0-stable, [3]1-stable; 
The string should conform to color syntax, such as "\code{red}" or "\code{#FF0000}".}

\item{LineColor}{a string that describes the color of polygonal boundary. 
\code{NA} means the boundary is not displayed.}

\item{Linewidth}{a numeric value, the line width of polygonal boundary.}
}
\value{
a \code{ggplot} object, show the percolation case.
}
\description{
This function visualizes the state distribution at each lattice 
point on square, triangular, or hexagonal lattices. Nodes are typically classified
into three categories: stable at "1", stable at "0", and oscillatory. The 
output of the function depends on the \code{ggplot} class.
}
