% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBoolBioNetwork.R
\name{LoadBoolBioNetwork}
\alias{LoadBoolBioNetwork}
\title{Load a Boolean genetic network}
\usage{
LoadBoolBioNetwork(
  NetName,
  NetType = c("BoolExpression", "TruthTable", "Threshold"),
  ThresMode = c("zero", "one", "self")
)
}
\arguments{
\item{NetName}{A string that denotes a valid file path or name for the target 
network.}

\item{NetType}{Which input format should be used? Options are restricted to 
\code{BoolExpression}, \code{TruthTable}, and \code{Threshold}.
\itemize{
\item \code{BoolExpression}: All genes in a single file are represented as Boolean expressions.
\preformatted{
  X_1 = X_3 AND X_4
  X_2 = NOT (X_1 OR X_3)
  ... ...
}
\item \code{TruthTable}: All genes stored in a specified directory, organized
according to the structural framework of \href{https://cellcollective.org}{Cell Collective Website}.
\preformatted{
  in /your_path/NetworkName/X_1.csv (for gene X_1)
     X_3, X_5, X_1
      0 ,  0 ,  0
      0 ,  1 ,  1
      1 ,  0 ,  1
      1 ,  1 ,  0
  in /your_path/NetworkName/X_2.csv (for gene X_2)
     X_4, X_2
      0 ,  0
      1 ,  1
      ... ...
}
\item \code{Threshold}: Each line represents one regulatory interaction in a 
 threshold network, where 1 denotes activation and 2 denotes inhibition.
\preformatted{
     X_1  X_3  1
     X_4  X_2  2
     X_5  X_2  1
      ... ...
}
}}

\item{ThresMode}{Threshold networks may exhibit cases where the sum of 
upstream inputs is zero, and different threshold patterns are defined as follows:
\itemize{
 \item{\code{zero}}: the controlled node is set to 0.
 \item{\code{one}}: the controlled node is set to 1.
 \item{\code{self}}: the controlled node retains its current state. In this 
case, each node is added a self-regulatory edge.
}
This parameter is effective only when the \code{NetType} parameter is \code{Threshold}.}
}
\value{
A four-element formatted list that records information of a Boolean network.
(See \link{BoolGRN_CellCollective})
}
\description{
Load genetic network information from external files.
}
