% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadManualNetwork.R
\name{LoadManualNetwork}
\alias{LoadManualNetwork}
\title{Load a user-defined network}
\usage{
LoadManualNetwork(
  Object,
  WhichCols = c(1, 2),
  BoolFunList = NULL,
  OBF_ratio = 0,
  OBF_type = "R"
)
}
\arguments{
\item{Object}{A formatted data to record regulatory patterns; support \code{matrix},
\code{data.frame}, \code{igraph} object (if available).}

\item{WhichCols}{An IntegerVector, specifies which columns should serve as 
the source and target columns (Default: \code{c(1,2)}).}

\item{BoolFunList}{A list in which each element is a Boolean function, with 
the list length equal to the system size. If \code{NULL}, the function will 
generate a list of Boolean functions randomly based on the parameters 
\code{OBF_ratio} and \code{OBF_type}. (Default: \code{NULL})}

\item{OBF_ratio}{A numeric value in \code{[0,1]} representing the proportion 
of ordered Boolean function; this value is invalid if \code{BoolFunList} is 
\code{FALSE}.}

\item{OBF_type}{A character denotes the type of ordered Boolean function, 
acceptable parameters are detailed in \link{BoolFun_Generator}.}
}
\value{
A four-element formatted list that records information of a Boolean 
network (See \link{BoolGRN_CellCollective}).
}
\description{
Instead of importing from files, this function enables users to 
define regulatory relationships and network structures to explore various 
system features.
}
