% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_ExampleFuns.R
\name{MulVFun_ExampleFuns}
\alias{MulVFun_ExampleFuns}
\title{three exemplified multi-valued functions}
\usage{
data(MulVFun_ExampleFuns)
}
\description{
three exemplified multi-valued functions
}
\details{
They are saved as data.frame. Call the variable directly to see the details.
\itemize{
  \item \code{Example_3v_LogicalFun}: a three-valued logical function
  \item \code{Example_5v_LogicalFun}: a five-valued logical function
  \item \code{Example_3v_Multiplexer}: a three-valued Multiplexer
}
Detail information, please see original paper 
[\href{https://doi.org/10.1007/s00500-007-0175-x}{Petrik2008}].
}
