% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_Generator.R
\name{MulVFun_Generator}
\alias{MulVFun_Generator}
\title{Generate a specific type of multi-valued function}
\usage{
MulVFun_Generator(
  MF_Type = c("R", "C", "D", "T"),
  k = 3L,
  L = 3L,
  Bias = NULL,
  CanaDeep = 1L,
  CanaVar = NULL,
  CanaVarNum = NULL,
  CanaLayerInfo = NULL,
  MappingTable = FALSE
)
}
\arguments{
\item{MF_Type}{A character included in the following:
 \itemize{
  \item \code{C}: Canalization
  \item \code{D}: Dominated-valued
  \item \code{T}: Linear threshold-based
  \item \code{R}: Random
}}

\item{k}{An integer that denotes the number of input variables (Default: 3).}

\item{L}{An integer that represents the level of multi-valued system (Default: 3).}

\item{Bias}{A NumericVector representing the probabilities of elements in 
non-controlled slots within mapping tables (See \code{Details}).
By default, \code{NULL}, which corresponds to the vector \code{rep(1.0/L, L)}.
If provided, the vector must have a length greater than L-1 and contain only
non-negative real numbers. The normalization of its first L elements determines
the baseline configuration.}

\item{CanaDeep}{An integer ranging from 1 to k, indicating k-layer canalization
for type \code{'C'}. (Default: 1)}

\item{CanaVar}{A \code{CanaDeep}-length non-repeating IntegerVector, indicating
which variables should be canalized, with values ranging from 1 to k.}

\item{CanaVarNum}{An integer vector of length \code{CanaDeep} specifying the 
number of each canalizing variable (ranging from 1 to L). If \code{CanaLayerInfo}
is appropriately provided, \code{CanaVarNum} is ignored. (Default: \code{NULL}, 
automatically set using \code{sample.int(L, CanaDeep, replace=TRUE)} implemented
in C++, rather than the native R function).}

\item{CanaLayerInfo}{A list containing detail information for canalization:
[[1]] for input (canalizing patterns); [[2]] for output (canalized patterns).
This argument must meet specific requirements (see \code{Details}). 
If not provided (Default: \code{NULL}), the function will automatically 
generate a random configuration. For detailed meaning, refer to \link{BoolFun_Type}.}

\item{MappingTable}{A logical value. Is the mapping table also returned? 
(Default: \code{FALSE})}
}
\value{
An integervector of length \eqn{L^k}{L^k}.
}
\description{
This function can generate following types: Canalization, 
Linear threshold-based, Dominated-valued.
}
\details{
The parameter Bias applies to non-controlled slots in the mapping table. 
For instance, a canalized function (k=3 and L=3) where \eqn{f(x_1=0)=2}{f(x1=0)=2}, indicates 
that only inputs with \eqn{x_1\not=0}{x1!=0} can be assigned probabilistic 
values according to the specified bias. 

The \code{CanaLayerInfo} must satisfy the following requirements: 
it must contain at least two sub-lists, representing the canalizing and 
canalized configurations, respectively. Each sub-list must have a length 
exceeding \code{k}. Within each sub-list, all canalizing and canalized values
must be strictly less than \code{L}. Furthermore, the elements in 
\code{CanaLayerInfo[[1]][[i]]} must be unique, which represent different 
canalizing values (See examples of \link{MulVFun_is_NestedCana}). 
Any violation of these conditions will lead to an error message and execution
halts.
}
\examples{
# Generate 3-input ternary function. Please note the canalizing/canalized 
# values in the following scenarios:
set.seed(1234L)
Example_01=MulVFun_Generator('C', 3L, 3L, MappingTable=TRUE)
Example_02=MulVFun_Generator('T', 3L, 3L, MappingTable=TRUE)
Example_03=MulVFun_Generator('D', 3L, 3L, MappingTable=TRUE)
# 1=Yes, -1=No (for Canalized)
# 1=Yes, 0=Unknown, -1=No (for Threshold and Domainted)

MulVFun_is_NestedCana(Example_01[,4],3L,3L)[[1]]
# 1
MulVFun_is_Threshold(Example_01[,4],3L,3L)[[1]]
# -1
MulVFun_is_Domainted(Example_01[,4],3L,3L)[[1]]
# -1
MulVFun_is_NestedCana(Example_02[,4],3L,3L)[[1]]
# -1
MulVFun_is_Threshold(Example_02[,4],3L,3L)[[1]]
# 1
MulVFun_is_Domainted(Example_02[,4],3L,3L)[[1]]
# -1
MulVFun_is_NestedCana(Example_03[,4],3L,3L)[[1]]
# -1
MulVFun_is_Threshold(Example_03[,4],3L,3L)[[1]]
# -1
MulVFun_is_Domainted(Example_03[,4],3L,3L)[[1]]
# 1

}
