% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_is_Signed.R
\name{MulVFun_is_Signed}
\alias{MulVFun_is_Signed}
\title{Decode the information of a multi-valued signed function}
\usage{
MulVFun_is_Signed(aVec, k, L, PrintOut = TRUE)
}
\arguments{
\item{aVec}{An IntegerVector that represents a mapping table of multi-valued 
function.}

\item{k}{An integer that denotes the number of input variables.}

\item{L}{An integer that represents the level of multi-valued system.}

\item{PrintOut}{A logical value. Should output be displayed in the terminal? 
(Default: \code{TRUE}).}
}
\value{
A list containing three elements:
\itemize{
\item \code{[[1]] IsSigned}: Is Signed? 1 for Yes, 0 for No.
\item \code{[[2]] Weight}: The weights of all pairwise relationships among
varibales; it is a L*(kL) matrix.
\item \code{[[3]] Delta}: The baselines of variables, ordered from low 
to high bits.}
}
\description{
The function analyzes and returns a list containing information
whether given function is signed, and corresponding weight matrix, and baselines.
It is also compatible with Boolean system (L=2).
}
\examples{
# Generate a domainted function and analyze it belong to signed class or not.
# set.seed(1001);
# A_MulV_Fun=MulVFun_Generator("D", k=3L, L=3L, MappingTable=TRUE);
# A_MulV_Fun_Sign=MulVFun_is_Signed(A_MulV_Fun[,4], 3, 3, TRUE);
# View(A_MulV_Fun); # Show the mapping table 
#       v2 v1 v0 f_out
#  [1,]  0  0  0     2
#  [2,]  0  0  1     2
#  [3,]  0  0  2     1
#  [4,]  0  1  0     2
#  [5,]  0  1  1     0
#  [6,]  0  1  2     0
#          ... ... 
# Check the "signed" regulatory patterns (User can test other cases)
# A_MulV_Fun_Sign[[2]]\%*\%c(1,0,0, 1,0,0, 1,0,0) ==> (-4,-1,0) ==> f_out(0,0,0)=2;
# A_MulV_Fun_Sign[[2]]\%*\%c(0,0,1, 1,0,0, 1,0,0) ==> (-2,-1,-5) ==> f_out(0,0,2)=1;
# A_MulV_Fun_Sign[[2]]\%*\%c(0,1,0, 0,0,1, 1,0,0) ==> (4,-1,1) ==> f_out(0,2,1)=0;
# ... ... 

set.seed(1001)
A_MulV_Fun=MulVFun_Generator("D", k=3L, L=3L, MappingTable=TRUE)
A_MulV_Fun_Sign=MulVFun_is_Signed(A_MulV_Fun[,4], 3, 3, TRUE)
A_MulV_Fun_Sign[[2]]\%*\%c(1,0,0, 1,0,0, 1,0,0)
A_MulV_Fun_Sign[[2]]\%*\%c(0,0,1, 1,0,0, 1,0,0)
A_MulV_Fun_Sign[[2]]\%*\%c(0,1,0, 0,0,1, 1,0,0)

}
