% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_QMForm.R
\name{MulVFun_QMForm}
\alias{MulVFun_QMForm}
\title{Show the Quine-McCluskey form of a multi-valued function}
\usage{
MulVFun_QMForm(
  aVec,
  k,
  L,
  VarsName = NULL,
  ShowType = c("print", "data.frame"),
  SourceTable = FALSE
)
}
\arguments{
\item{aVec}{An IntegerVector that represents a mapping table of multi-valued function.}

\item{k}{An integer that denotes the number of input variables.}

\item{L}{An integer that represents the level of multi-valued system.}

\item{VarsName}{A CharacterVector that represents each name of variable.}

\item{ShowType}{Which output format should be used? Options are restricted to
\code{print} (printed in terminal) and \code{data.frame} (in data.frame form).}

\item{SourceTable}{A logical value. Should return the source mapping table? 
It is ignored when \code{ShowType} is \code{print} (Default: \code{FALSE}).}
}
\value{
Three different cases:
\itemize{
   \item \code{NULL}: Only print in terminal (\code{ShowType=="print"})
   \item \code{data.frame}: \code{ShowType} is \code{"data.frame"} and 
\code{SourceTable} is \code{FALSE}.
   \item \code{list}: \code{ShowType} is \code{"data.frame"} and 
\code{SourceTable} is \code{TRUE}. [[1]] is the result; [[2]] is the 
source funciton.}
}
\description{
The function employs our native C++ standard multi-valued 
Quine-McCluskey algorithm to obtain prime implicants. Because of algorithms 
and system properties, the result is not unique for multi-valued system. See reference 
[\href{https://doi.org/10.1007/s00500-007-0175-x}{Petrik2008}].
}
\examples{
# Calculate the Quine-McCluskey form of multi-valued function.
# Analyze a three-valued multiplexer: 
#             | a, iff x=0
# f(x,c,b,a)= | b, iff x=1
#             | c, iff x=2
# Example_3v_Multiplexer; # Show the multiplexer.
# MulVFun_QMForm(Example_3v_Multiplexer[,5], k=4, L=3,
#   VarsName=colnames(Example_3v_Multiplexer)[4:1]);
# 
# Return its QMC form:
# x  c  b  a f(*)
# 2  1 NA NA    1
# 2  2 NA NA    2
# 1 NA  1 NA    1
# 1 NA  2 NA    2
# 0 NA NA  1    1
# 0 NA NA  2    2
MulVFun_QMForm(Example_3v_Multiplexer[,5], k=4, L=3, 
   VarsName=colnames(Example_3v_Multiplexer)[4:1])

}
