% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_Recovery.R
\name{MulVFun_Recovery}
\alias{MulVFun_Recovery}
\title{Recover a multi-valued function from multi-valued model}
\usage{
MulVFun_Recovery(aMulVNet, GeneName, MaxValue = 2L, MapTab = FALSE)
}
\arguments{
\item{aMulVNet}{A four-element formatted list that records information of a multi-valued}

\item{GeneName}{A string that denotes the gene name.}

\item{MaxValue}{An integer that represents the max value of the gene (Default: 2).}

\item{MapTab}{A logical value. Is the mapping table also returned? 
(Default: \code{FALSE})}
}
\value{
An IntegerVector or a mapping table.
}
\description{
Unlike the standard multi-valued logical paradigm, 
we adopt the definition of so-called qualitative networks; please refer to 
[\href{https://doi.org/10.1186/1752-0509-1-4}{Schaub2007}]
}
