% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_Sensitivity.R
\name{MulVFun_Sensitivity}
\alias{MulVFun_Sensitivity}
\title{Calculate the sensitivity of a multi-valued function}
\usage{
MulVFun_Sensitivity(Map_Vec, k, L)
}
\arguments{
\item{Map_Vec}{An IntegerVector that represents a mapping table of 
multi-valued function.}

\item{k}{An integer that denotes the number of input variables.}

\item{L}{An integer that represents the level of multi-valued system.}
}
\value{
A numeric value representing the sensitivity of the given function.
}
\description{
Correspodning concepts of multi-valued systems are generalized 
from the Boolean system.
}
\examples{
# Show the canalized structure of a multi-valued nested canalized function.
# MulVFun_Sensitivity(c(0,1,2, 1,0,1, 2,2,0), k=2, L=3);
# Return 1.777778

MulVFun_Sensitivity(c(0,1,2, 1,0,1, 2,2,0), k=2, L=3)

}
