% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_CheckValidBoolFun.R
\name{r_CheckValidBoolFun}
\alias{r_CheckValidBoolFun}
\title{Determines whether the input vector is a valid mapping result.}
\usage{
r_CheckValidBoolFun(Bit_Vec)
}
\arguments{
\item{Bit_Vec}{a bool vector, represents a truth table of Boolean function.}
}
\value{
a List, [[1]] Boolean vector, [[2]] in-degree
}
\description{
Determines whether the input vector is a valid mapping result.
}
