% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_CheckValidMulVFun.R
\name{r_CheckValidMulVFun}
\alias{r_CheckValidMulVFun}
\title{Multi-valued function checker}
\usage{
r_CheckValidMulVFun(a_Vec, k, L)
}
\arguments{
\item{a_Vec}{An IntegerVector representing a multi-valued function.}

\item{k}{An integer that denotes the number of input variables.}

\item{L}{An integer that represents the level of multi-valued system.}
}
\value{
IntegerVector or prints an error message and execution halts.
}
\description{
Check whether the input vector is a valid multi-valued function
}
