% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_GenerateTruthTable.R
\name{r_GenerateTruthTable}
\alias{r_GenerateTruthTable}
\title{Generate the header of truth table}
\usage{
r_GenerateTruthTable(inDegree = 2L, SpinLike = FALSE)
}
\arguments{
\item{inDegree}{An integer that denotes the number of input variables.}

\item{SpinLike}{A logical value. Should output the spin-like form?}
}
\value{
A data.frame of input case enumeration
}
\description{
Generate state enumeration table for the given number of input variables
}
