% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_GenerateTruthTable_M.R
\name{r_GenerateTruthTable_M}
\alias{r_GenerateTruthTable_M}
\title{Generate the header of mapping table}
\usage{
r_GenerateTruthTable_M(inDegree = 2L, Lsystem = 2L)
}
\arguments{
\item{inDegree}{An integer that denotes the number of input variables (Default: 2).}

\item{Lsystem}{An integer that represents the level of multi-valued system (Default: 2).}
}
\value{
a data.frame of input case enumeration
}
\description{
Generate state enumeration table for the given number of input 
variables (multi-valued version)
}
