% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_UnfreezeInputNode.R
\name{r_UnfreezeInputNode}
\alias{r_UnfreezeInputNode}
\title{Treat all input/external nodes as self-loop node.}
\usage{
r_UnfreezeInputNode(RealBioNet)
}
\arguments{
\item{RealBioNet}{List[[4]], a formatted data, record biological network information}
}
\value{
List[[4]], a formatted data, record biological network information
}
\description{
Treat all input/external nodes as self-loop node.
}
