% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_BoolFun.R
\name{BoolBioNet_BoolFun}
\alias{BoolBioNet_BoolFun}
\title{Explore features of Boolean functions in biological/genetic networks}
\usage{
BoolBioNet_BoolFun(OneBioNet, Type = c("rough", "detail"))
}
\arguments{
\item{OneBioNet}{a 4-element formatted list, record Boolean network information.}

\item{Type}{Which statistical method is used (\code{rough} or \code{detail})}
}
\value{
\code{data.frame}, counts the number of various types of functions (\code{Type} is \code{rough}).
\code{data.frame}, records the specific cases of each function within systems (\code{Type} is \code{detail}).
}
\description{
This function can calculate and return the categories, sensitivity, effectiveness, and complexity of each Boolean function in a given Boolean network.
}
\details{
Codes can identify function types, including canalized, threshold, clique,
dominating, effective, monotone, signed, spin-like. The definations of various funciton can be seen 
the document of \link{BoolFun_Type}. Please note that single regulatory and external factor are 
ignored from statistics (only 2 <= in-degree <= 16, other cases return \code{NA}). When \code{Type} is 
\code{rough}, function returns a \code{data.frame} that contains statistics of each "in-degree" cluster. 
When \code{Type} is \code{detail}, the returned data.frame contains each case of Boolean function.
}
\examples{
# Analysis of one real Boolean genetic network.
# BoolBioNet_BoolFun(BoolGRN_CellCollective$c_1778,"rough");
# Return a data.frame:
#                 Number Canalized Clique Monotone Signed Threshold
# InDegree_1           1        NA     NA       NA     NA        NA
# InDegree_2           3         3      3        1      3         3
# InDegree_3           3         3      2        0      3         3
#                             ... ...
BoolBioNet_BoolFun(BoolGRN_CellCollective$c_1778,"rough")

# BoolBioNet_BoolFun(BoolGRN_CellCollective$c_1778,"detail");
# Return a data.frame
#            Input Canalized Clique Monotone Signed Threshold Sensitivity Effectiveness Complexity
# ADD            5      TRUE  FALSE    FALSE   TRUE      TRUE   1.3125000      1.802083        3.0
# ATM            3      TRUE  FALSE    FALSE   TRUE      TRUE   1.2500000      1.562500        2.0
# ATR            4      TRUE  FALSE    FALSE   TRUE      TRUE   1.2500000      1.625000        2.5
#                                ... ...
BoolBioNet_BoolFun(BoolGRN_CellCollective$c_1778,"detail")

}
