% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolFun_Sensitivity.R
\name{BoolFun_Sensitivity}
\alias{BoolFun_Sensitivity}
\title{Calculate the sensitivity of a Boolean function}
\usage{
BoolFun_Sensitivity(Bit_Vec)
}
\arguments{
\item{Bit_Vec}{a bool (or 0/1-integral) vector, represents a truth table of Boolean function.}
}
\value{
double, sensitivity of a Boolean function
}
\description{
The analysis depends on the enumeration of input vectors and 
observe its sensitivity of one-bit perturbations, such as f(..., x_i=0, ...)
is identical to f(..., x_i=1, ...) or not in all possible inputs. Detail 
concepts can see the paper [\href{https://doi.org/10.1103/PhysRevLett.93.048701}{Shmulevich,2004}].
}
\examples{
# Calculate the sensitivity of a Boolean function \code{c(1,1,0,1,1,1,0,0)}.
# BoolFun_Sensitivity(c(1,1,0,1,1,1,0,0))
# Return 1.25000
BoolFun_Sensitivity(c(1,1,0,1,1,1,0,0))

}
