% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_Complexity}
\alias{c_BF_Complexity}
\title{C++ prototype function for BoolFun_Complexity()}
\usage{
c_BF_Complexity(boolfunss, k)
}
\arguments{
\item{boolfunss, }{bool function logical vector.}

\item{k, }{input variable size.}
}
\description{
C++ prototype function for BoolFun_Complexity()
}
\details{
Please see document of \link{BoolFun_Complexity}
}
