% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_EffectiveEdges}
\alias{c_BF_EffectiveEdges}
\title{C++ prototype function for BoolFun_EffectiveEdges()}
\usage{
c_BF_EffectiveEdges(boolfunss, k)
}
\arguments{
\item{boolfunss, }{truth table.}

\item{k, }{input variable size.}
}
\description{
C++ prototype function for BoolFun_EffectiveEdges()
}
\details{
Please see document of \link{BoolFun_EffectiveEdges}
}
