% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_FrameTruthTable}
\alias{c_FrameTruthTable}
\title{C++ prototype function for FrameTruthTable()}
\usage{
c_FrameTruthTable(VarNum)
}
\arguments{
\item{VarNum, }{input variable size.}
}
\description{
C++ prototype function for FrameTruthTable()
}
