% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulF_QuineMcCluskey}
\alias{c_MulF_QuineMcCluskey}
\title{C++ prototype function for MulVFun_QMForm()}
\usage{
c_MulF_QuineMcCluskey(avec, k, L)
}
\arguments{
\item{avec, }{IntegerVector, each element in {0,1,...L-1}}

\item{k, }{input number (in-degree)}

\item{L, }{system of L:{v0,v1,...v_{L-1}}}
}
\value{
a matrix, rows are multi-valued prime implicants, columns are
each variable's state [0~L-1], -1 means "*". Last column is mapped results.
}
\description{
C++ prototype function for MulVFun_QMForm()
}
\details{
Please see document of \link{MulVFun_QMForm}
}
