% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigLatticePlot.R
\name{FigLatticePlot}
\alias{FigLatticePlot}
\title{Show the percolation}
\usage{
FigLatticePlot(
  Values = NULL,
  Length = NULL,
  Height = NULL,
  Type = c(4L, 3L, 6L),
  ColorFills = c("#efefef", "#EE312E", "#2156A6"),
  LineColor = "#111111",
  Linewidth = 0.7
)
}
\arguments{
\item{Values}{integer vector, a valid file/address of target GenNet.}

\item{Length}{an integer, Length of lattice matrix.}

\item{Height}{an integer, Height of lattice matrix.}

\item{Type}{an integer, 4, Square; 3, triangle; 6, hexagon.}

\item{ColorFills}{character vector [3], [1] unstable, [2]0-stable, [3]1-stable}

\item{LineColor}{a character, color of polygonal boundary. (\code{NA}, no diaply)}

\item{Linewidth}{numeric value, line width of polygonal boundary.}
}
\value{
a \code{ggplot} class, show the percolation.
}
\description{
This function displays the state distribution at each point under 
square/triangle/hexagon lattice. Nodes are generally classified into three 
categories: stable-1, stable-0, and oscillatory. This function depends on the 
frame of \code{ggplot} class.
}
