% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadManualNetwork.R
\name{LoadManualNetwork}
\alias{LoadManualNetwork}
\title{Load an user-defined network}
\usage{
LoadManualNetwork(
  Object,
  WhichCols = c(1, 2),
  BoolFunList = NULL,
  OBF_ratio = 0,
  OBF_type = "R"
)
}
\arguments{
\item{Object}{a formatted data, record regulatory patterns (Support \code{matrix}, \code{data.frame}, \code{igraph-object})}

\item{WhichCols}{integer Vector, Which columns should be source and target columns (Default: \code{c(1,2)})}

\item{BoolFunList}{list, Boolean function, its length equals to size of system, and each element denotes a Boolean function. (Default: \code{NULL})}

\item{OBF_ratio}{a numeric in \code{[0,1]}, proportion of ordered  Boolean function; it will be invalid if \code{is.null(BoolFunList)} is \code{FALSE}.}

\item{OBF_type}{a character, type of ordered Boolean function, acceptable parameters can be found document of \link{BoolFun_Generator}.}
}
\value{
a list contains member names of system, topological structure, and Boolean function.
}
\description{
Instead of importing from files, this function allows users to 
define regulatory relationships and network structures to explore various 
system features.
}
