% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_ExampleFuns.R
\name{MulVFun_ExampleFuns}
\alias{MulVFun_ExampleFuns}
\title{3 exemplified multi-valued functions}
\usage{
data(MulVFun_ExampleFuns)
}
\description{
3 exemplified multi-valued functions
}
\details{
They are saved as data.frame. Call the variable directly to see the details.
\itemize{
  \item \code{five_valued_LogicalFun}:  a five-valued logical function
  \item \code{three_valued_LogicalFun}: a three-valued logical function
  \item \code{three_valued_Multiplexer}: a three-valued Multiplexer ["MulVFun_ExampleFuns"]
}
Detail information, please see original paper \href{https://doi.org/10.1007/s00500-007-0175-x}{Petrik2008}
}
