% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_Sensitivity.R
\name{MulVFun_Sensitivity}
\alias{MulVFun_Sensitivity}
\title{Calculate the sensitivity of a multi-valued function}
\usage{
MulVFun_Sensitivity(Map_Vec, k, L)
}
\arguments{
\item{Map_Vec}{a multi-valued vector, represents a mapping table of multi-
valued functions.}

\item{k}{An integer for the number of input-variable.}

\item{L}{An integer for the level of discrete system.}
}
\value{
double, sensitivity of a multi-valued function.
}
\description{
Correspodning concepts of multi-valued systems are generalized 
from the Boolean system.
}
\examples{
# Show the canalized structure of a multi-valued nested canalized function.
# MulVFun_Sensitivity(c(0,1,2, 1,0,1, 2,2,0), k=2, L=3)
# Return 1.777778
MulVFun_Sensitivity(c(0,1,2, 1,0,1, 2,2,0), k=2, L=3)

}
